+++ R/draw-panel.R
@@ -1,5 +1,3 @@
-
-
 pretty_format_numbers <- function(labels) {
   n_decimals <-
     max(sapply(stringr::str_split(
@@ -285,6 +283,38 @@ as.barplot.x <- function(bp.data, x, xlim, bar.stacked, log_scale) {
   }
 }
 
+drawwaterfall <- function(l, data, xlim, ylim, log_scale, horiz) {
+  snames <- sapply(data$series, function(s) s$name)
+  data$bars <- NULL # remove the pre-fetched bardata (used for things like setting ylimits)
+  bars <- extract_bar_data(data, "waterfall")$bars
+
+  if (ncol(bars$bardata) > 0 && nrow(bars$bardata) > 0) {
+    for (i in 1:nrow(bars$bardata)) {
+      draw_data <- bars$bardata
+      draw_data[-i,] <- NA
+
+      if (i > 1 && i < nrow(bars$bardata)) {
+        y_offset <- sum(bars$bardata[1:(i-1),], na.rm = TRUE)
+      } else {
+        y_offset <- 0
+      }
+
+      drawbars_(
+        l,
+        data,
+        draw_data,
+        bars$colours,
+        bars$bordercol,
+        xlim,
+        ylim - y_offset,
+        TRUE,
+        log_scale,
+        horiz
+      )
+    }
+  }
+}
+
 drawbar <- function(l, bardata, colours, bordercol, xlim, ylim, bar.stacked, log_scale, horiz) {
   if (any(bardata != 0)) { # otherwise no point - and doing so causes errors with log scale plots
     graphics::par(mfg = l)
@@ -305,12 +335,7 @@ drawbar <- function(l, bardata, colours, bordercol, xlim, ylim, bar.stacked, log
   }
 }
 
-drawbars <- function(l, data, xlim, ylim, bar.stacked, log_scale, horiz) {
-  out <- get_bar_data(data)
-  bardata <- out$bardata
-  colours <- out$colours
-  bordercol <- out$bordercol
-
+drawbars_ <- function(l, data, bardata, colours, bordercol, xlim, ylim, bar.stacked, log_scale, horiz) {
   if (ncol(bardata) > 0) {
     out <- convert_to_plot_bardata(bardata, data)
     bardata_p <- out$p
@@ -330,6 +355,11 @@ drawbars <- function(l, data, xlim, ylim, bar.stacked, log_scale, horiz) {
       drawbar(l, bardata_n, colours, bordercol, ylim, xlim, bar.stacked, rotate_log_scale(log_scale), horiz)
     }
   }
+}
+
+drawbars <- function(l, data, xlim, ylim, bar.stacked, log_scale, horiz) {
+  out <- get_bar_data(data)
+  drawbars_(l, data, out$bardata, out$colours, out$bordercol, xlim, ylim, bar.stacked, log_scale, horiz)
 }
 
 draw_blankplot <- function(l, xlim, ylim, log_scale, horiz) {
@@ -370,6 +400,7 @@ drawpanel <- function(p, data, shading, bgshadings, margins, layout, yunits, xun
   gridsandborders(p, layout, yunits, xunits, yticks, xlabels, ylim, xlim, dropxlabel, srt, data$ts)
 
   if (!is_empty(data)) drawbars(l, data, xlim, ylim, bar.stacked, log_scale, layout == "1h")
+  if (!is_empty(data)) drawwaterfall(l, data, xlim, ylim, log_scale, layout == "1h")
 
   # Reset the plot after the bars (which use different axis limits), otherwise lines and shading occur in the wrong spot
   draw_blankplot(l, xlim, ylim, log_scale, layout == "1h")