+++ tests/testthat/test-gg.R
@@ -52,6 +52,41 @@ test_that("Layers", {
   # More complex step
   foo <- arphitgg(data) + agg_step(aes = agg_aes(x = date, y = unemployment, group = state))
   expect_true(check_graph(foo, "gg-layer-step"))
+
+  # Waterfall graph
+  data  <- data.frame(x = letters[1:6], y = c(2,1.1,-0.5,-0.2,0.4,2.8))
+  foo <- arphitgg(data) + agg_waterfall(agg_aes(x=x,y=y))
+  expect_true(check_graph(foo, "gg-layer-waterfall"))
+
+  # Below the axis
+  data  <- data.frame(x = letters[1:6], y = -c(2,1,-0.5,-0.2,0.4,2.7))
+  foo <- arphitgg(data) + agg_waterfall(agg_aes(x=x,y=y))
+  expect_true(check_graph(foo, "gg-layer-waterfall-below"))
+
+  # Cross the axis
+  data <- data.frame(x = letters[1:4],
+                     y = c(0.5, -1, 0.2, -0.3))
+  foo <- arphitgg(data) + agg_waterfall(agg_aes(x,y))
+  expect_true(check_graph(foo, "gg-layer-waterfall-crossing"))
+
+  # Only positive changes
+  data  <- data.frame(x = letters[1:4], y = c(2,1,0.5,3.5))
+  foo <- arphitgg(data) + agg_waterfall(agg_aes(x=x,y=y))
+  expect_true(check_graph(foo, "gg-layer-waterfall-positive-changes"))
+
+  # Waterfall with groups
+  data <- data.frame(x = c('start','a','a','b','b','end'),
+                     y = c(1, 0.5, -0.4, 0.2, 0.1, 1.4),
+                     group = c(1, 2, 3, 2, 3, 4),
+                     order = c(1,2,2,3,3,4))
+  foo <- arphitgg(data) +
+    agg_waterfall(agg_aes(x=x,y=y,group=group,order=order))
+  expect_true(check_graph(foo, "gg-layer-waterfall-groups"))
+
+  # Waterfall mixed with other layers
+  data <- data.frame(x=letters[1:6],y=c(1,0.5,-0.2,0.2,-0.5,1),z=1:6)
+  foo <- arphitgg(data) + agg_waterfall(agg_aes(x,y,x))+agg_point(agg_aes(x,z))
+  expect_true(check_graph(foo, "gg-layer-waterfall-with-point"))
 })
 
 test_that("Aesthetic inheritance", {